refactor(agents): `create agent` -> `connect agent`
